requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0